Fix 4-way deadlock
To help track down a deadlock, added context logging to lock/unlock
calls in kvraft server. The deadlock is described as the 'The four-way
deadlock' in https://thesquareplanet.com/blog/students-guide-to-raft/.

Fixed by taking the kv's lock after making the call to rf.Start().

TestConcurrent passes for a small number (<4) of clients.